renamed instalkergram to crud, and create stagram routes
