+++ app.js
@@ -10,10 +10,12 @@ app.use(express.static('public')); //Serving front from public
 app.use(cors());
 
 // routes
-app.post('/create-category', routes.instalkergram.createCategory);
-app.post('/update-category', routes.instalkergram.updateCategory);
-app.get('/get-images/:category', routes.instalkergram.getImages);
-app.delete('/delete-image', routes.instalkergram.deleteImage);
+app.post('/create-category', routes.crud.createCategory);
+app.post('/update-category', routes.crud.updateCategory);
+app.get('/get-images/:category', routes.crud.getImages);
+app.delete('/delete-image', routes.crud.deleteImage);
+
+app.get('/getUrlsFromInstagram/:token', routes.instagram.getUrls);
 
 app.listen(3000, function () {
   console.log('rodando na porta 3000 :D!');
