+++ package.json
@@ -22,6 +22,8 @@
     "express": "^4.14.0",
     "mkdirp": "^0.5.1",
     "mongodb": "^2.2.5",
+    "request": "^2.74.0",
+    "request-promise": "^4.1.0",
     "underscore": "^1.8.3"
   },
   "homepage": "https://github.com/Sauloxd/instalkergram#readme"
